requirements.txt
absl-py>=2.1.0
aiohappyeyeballs>=2.4.3
aiohttp>=3.10.10
aiohttp-cors>=0.7.0
aiosignal>=1.3.1
annotated-types>=0.7.0
anyio>=4.6.2.post1
astunparse>=1.6.3
attrs>=24.2.0
bioimageio.core>=0.6.10
bioimageio.spec>=0.5.3.3
cachetools>=5.5.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
click>=8.1.7
cloudpickle>=3.1.0
coloredlogs>=15.0.1
colorful>=0.5.6
conda-pack>=0.8.0
coverage>=7.6.4
Deprecated>=1.2.14
distlib>=0.3.9
distro>=1.9.0
dm-tree>=0.1.8
dnspython>=2.7.0
email_validator>=2.2.0
Farama-Notifications>=0.0.4
fastapi>=0.115.4
fastremap>=1.15.0
fastrlock>=0.8.2
filelock>=3.16.1
flatbuffers>=24.3.25
frozenlist>=1.5.0
fsspec>=2024.10.0
gast>=0.6.0
google-api-core>=2.21.0
google-auth>=2.35.0
google-pasta>=0.2.0
googleapis-common-protos>=1.65.0
grpcio>=1.67.0
gymnasium>=1.0.0
h11>=0.14.0
h5py>=3.12.1
httptools>=0.6.4
humanfriendly>=10.0
hypha_rpc>=0.20.38
idna>=3.10
imagecodecs>=2024.9.22
imageio>=2.36.0
importlib_metadata>=8.5.0
iniconfig>=2.0.0
jax-jumpy>=1.0.0
Jinja2>=3.1.4
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
keras>=3.6.0
lazy_loader>=0.4
libclang>=18.1.1
linkify-it-py>=2.0.3
llvmlite>=0.43.0
loguru>=0.7.2
lz4>=4.3.3
Markdown>=3.7
markdown-it-py>=3.0.0
MarkupSafe>=3.0.2
mdit-py-plugins>=0.4.2
mdurl>=0.1.2
memray>=1.14.0
ml-dtypes>=0.4.1
mpmath>=1.3.0
msgpack>=1.1.0
multidict>=6.1.0
munch>=4.0.0
namex>=0.0.8
natsort>=8.4.0
numba>=0.60.0
numpy>=2.0.2
onnxruntime>=1.19.2
opencensus>=0.11.4
opencensus-context>=0.1.3
opt_einsum>=3.4.0
optree>=0.13.0
packaging>=24.1
pandas>=2.2.3
pillow>=11.0.0
pip-review>=1.3.0
platformdirs>=4.3.6
pluggy>=1.5.0
pooch>=1.8.2
prometheus_client>=0.21.0
propcache>=0.2.0
proto-plus>=1.25.0
protobuf>=5.28.3
psutil>=6.1.0
py-spy>=0.3.14
pyarrow>=18.0.0
pyasn1>=0.6.1
pyasn1_modules>=0.4.1
pydantic>=2.9.2
pydantic-settings>=2.6.0
pydantic_core>=2.23.4
Pygments>=2.18.0
PyJWT>=2.9.0
pytest>=8.3.3
pytest-asyncio>=0.24.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
pytz>=2024.2
PyYAML>=6.0.2
referencing>=0.35.1
requests>=2.32.3
rich>=13.9.3
roifile>=2024.9.15
rpds-py>=0.20.0
rsa>=4.9
ruyaml>=0.91.0
scikit-image>=0.24.0
scipy>=1.13.0
setuptools>=75.2.0
sh>=2.1.0
shellingham>=1.5.4
shortuuid>=1.0.13
six>=1.16.0
smart-open>=7.0.5
sniffio>=1.3.1
starlette>=0.41.2
sympy>=1.13.1
tensorboard>=2.18.0
tensorboard-data-server>=0.7.2
tensorboardX>=2.6.2.2
tensorflow>=2.18.0
termcolor>=2.5.0
textual>=0.85.1
torch>=2.5.0
tqdm>=4.66.5
triton>=3.1.0
typer>=0.12.5
typing_extensions>=4.12.2
tzdata>=2024.2
uc-micro-py>=1.0.3
urllib3>=2.2.3
uvicorn>=0.32.0
uvloop>=0.21.0
virtualenv>=20.27.0
watchfiles>=0.24.0
websockets>=13.1
Werkzeug>=3.1.0
wheel>=0.44.0
yarl>=1.16.0
zipp>=3.20.2
